WIP: adding virtual object serialization method to entities
